Task40

+++ Task43/Program.cs
@@ -4,13 +4,13 @@
 значения b1, k1, b2 и k2 задаются пользователем.*/
 
 Console.WriteLine ("Введите значение k1: ");
-int yk1 = Convert.ToInt32(Console.ReadLine());
+float yk1 = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine ("Введите значение b1: ");
-int yb1 = Convert.ToInt32(Console.ReadLine());
+float yb1 = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine ("Введите значение k2: ");
-int yk2 = Convert.ToInt32(Console.ReadLine());
+float yk2 = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine ("Введите значение b2: ");
-int yb2 = Convert.ToInt32(Console.ReadLine());
+float yb2 = Convert.ToInt32(Console.ReadLine());
 
 if (yk1 == yk2 & yb1 == yb2){
     Console.WriteLine("При указанных k1, b1 и k2, b2 прямые будут совпадать!");
@@ -19,8 +19,8 @@ if (yk1 == yk2 & yb1 == yb2){
     Console.WriteLine("При указанных k1 и k2 прямые не будут иметь точек пересечения, прямые параллельны!");
 } else{
     // Нахождения ОХ точки пересечения:
-    int ox = (yb2 - yb1) / (yk1 - yk2);
+    float ox = (yb2 - yb1) / (yk1 - yk2);
     // Нахождения ОY точки пересечения:
-    int oy = (yk1 * yb2 - yk2 * yb1) / (yk1 - yk2);
+    float oy = (yk1 * yb2 - yk2 * yb1) / (yk1 - yk2);
     Console.WriteLine($"Точка пересечения имеет координату -> ({ox}, {oy})");
 }
